+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Added
 - Python 3 support
 - --all-templates option for `htmd start`
+- templates command
 
 ## [1.5.0] - 2015-02-16
 ### Added